Introduce new API also on macOS
Support to produce frames encoded in JSON and in the
binary format is now available also for macOS.

Actual JSON / binary format dump functions moved in
a separate file.
